requirements.txt
pywinauto==0.6.8
opencv-python==4.5.5.64
pillow==9.0.1
pytest==7.0.1
Appium-Python-Client==2.1.0
selenium==4.1.3 